employer/models.py
from django.db import models
from datetime import *
import uuid
from django.db import models
from django.contrib.auth.models import User
from django.dispatch import receiver
from ckeditor.fields import RichTextField
from phone_field import PhoneField
from django.db.models.signals import post_save, pre_save
from ckeditor.fields import RichTextField
import datetime
