requirements.txt
arrow==1.2.3
asgiref==3.7.1
async-timeout==4.0.2
binaryornot==0.4.4
bleach==6.0.0
boto3==1.26.140
botocore==1.29.140
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
coverage==7.2.7
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-dotenv==1.4.2
django-extensions==3.2.1
django-filter==23.2
django-js-asset==2.0.0
django-phone-field==1.8.1
django-redis==5.2.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.3.0
django-sslserver==0.22
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
facebook-sdk==3.1.0
filelock==3.12.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
jwcrypto==1.4.2
MarkupSafe==2.1.3
msgpack==1.0.5
oauthlib==3.2.2
pipenv==2023.6.2
platformdirs==3.5.1
protobuf==4.23.1
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-jose==3.3.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
ratelim==0.1.6
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==5.4.0
text-unidecode==1.3
typing_extensions==4.6.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.0
virtualenv-clone==0.5.7
webencodings==0.5.1
whitenoise==6.4.0
wrapt==1.15.0
